kernel_vpp: delete VPP config from previous strongswan process
On startup, set the IPsec "originator" identifier in VPP and delete any
existing SAs and/or policies in VPP with that originator value.

On subsequent connections to a VPP process, clear the cached information
about policies, SAs and routes installed in a previous VPP process.
This will force kernel_vpp to create all VPP config from scratch.