kernel-vpp: place lower bound on read timeout option
VPP's svm_queue_sub() function takes a timeout parameter,
but even though the underlying svm_queue_timedwait_inline() function
accepts a duration as floating point, which can be less than 1, the
timeout parameter to svm_queue_sub() is an integer and any fractional
part of the value is truncated.

vac_rx_thread_fn() takes the read timeout and divides it in half.  If
the configured timeout is less than two, the effective timeout becomes
zero and vac_rx_thread_fn ends up in a tight loop.

Check the timeout during plugin initialization and apply a minimum.